+++ src/main/java/dev/akorovai/backend/review/response/ReviewResponse.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @Builder
 public class ReviewResponse {
     private Long id;
-    private Short rating;
+    private Integer rating;
     private String comment;
     private LocalDateTime createdAt;
     private String subject;
